+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/animation/RotationFunctionStatelessAnimation.kt
@@ -28,6 +28,6 @@ class RotationFunctionStatelessAnimation<T : Entity>(
 ) : StatelessAnimation<T, ModelFrame>(frame) {
     override val targetFrame = ModelFrame::class.java
     override fun setAngles(entity: T?, model: PoseableEntityModel<T>, state: PoseableEntityState<T>?, limbSwing: Float, limbSwingAmount: Float, ageInTicks: Float, headYaw: Float, headPitch: Float) {
-        part.addRotation(axis, function(timeVariable(entity?.let { model.getState(it) }, limbSwing, ageInTicks) ?: 0F))
+        part.addRotation(axis, function(timeVariable(state, limbSwing, ageInTicks) ?: 0F))
     }
 }